pyproject.toml
[project]
name = "anansi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic>=1.4.1",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.9",
    "langsmith>=0.2.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
]
